refactored parse functions to own util file
